+++ frontend/src/app/(dashboard)/materiais/materiais.routes.ts
@@ -3,6 +3,7 @@ import { Routes } from '@angular/router';
 export const materiaisRoutes: Routes = [
   {
     path: '',
+    title: 'Materiais',
     loadComponent: () =>
       import('./materiais-home/materiais-home.component').then(
         (m) => m.MateriaisHomeComponent,
@@ -10,6 +11,7 @@ export const materiaisRoutes: Routes = [
   },
   {
     path: ':categoriaSlug',
+    title: 'Materiais',
     loadComponent: () =>
       import('./material-categoria/material-categoria.component').then(
         (m) => m.MaterialCategoriaComponent,